Make frame walk correct when called outside of a class
Mongothon initialization was failing when called outside a class (for instance from an ipython notebook). This is because you don't have a containing class from where you're calling. Added a simple test to see if the getmodule() call returns None, returning None if so. The generated model class won't have a module in this case, but this is not a problem in most cases.